feat(Argo CD) : Add Argo CD deployment and tools
To install Argo CD in your minikube, use install_argocd_in_k8s.sh
This script will create a new namespace "argocd" and deploy all
the needed components. Finally, it display the username and password
to login the Argo CD (UI or command line).

NOTE:
1. You need to install Argo CD CLI to run the next script.
2. You need to login by the given credencials.

To create the inventory application, use create_argocd_inventory_app.sh
This script deploy the inventory-app with all the needed k8s files. In
the end of the run, it display the application list and the details
about the inventory application.

The last file is the YAML file of the inventory-app deployment.

Issue: #8

Signed-off-by: Yaakov Neuman <yakinew@yahoo.com>