feat(Jenkins) : update flask-app image version
When the Jenkins pipeline finish and create new image in Docker Hub,
we would like it to update the version of the flask application to
use the new image. Doing that will cause the ArgoCD to update the
k8s cluster automatcally.

NOTE: The commit MUST contain the string [ci skip]. Otherwise it will
trigger the Jenkins pipeline again with endless loop.

Issue: #16

Signed-off-by: Yaakov Neuman <yakinew@yahoo.com>